+++ mycodo/databases/mycodo_db/__init__.py
@@ -31,11 +31,17 @@ from .models_5 import (
     DisplayOrder,
     Method,
     Misc,
-    CameraTimelapse,
-    CameraStream,
-    CameraStill,
+    Camera,
+    CameraType,
+    Remote,
+    Role,
     SMTP,
-    Remote
+    User
+)
+
+from mycodo.config import (
+    CAM_TYPES,
+    USER_ROLES
 )
 
 logger = logging.getLogger(__name__)
@@ -95,11 +101,12 @@ def populate_db(db_path):
     Session = sessionmaker(bind=engine)
     session = Session()
     try:
+        for index, each_camera_type in enumerate(CAM_TYPES, 1):
+            insert_or_ignore(Camera(id=index, camera_type=each_camera_type), session)
+        for each_role in USER_ROLES:
+            insert_or_ignore(Role(**each_role), session)
         insert_or_ignore(AlembicVersion(), session)
         insert_or_ignore(DisplayOrder(id=1), session)
-        insert_or_ignore(CameraTimelapse(id=1), session)
-        insert_or_ignore(CameraStill(id=1), session)
-        insert_or_ignore(CameraStream(id=1), session)
         insert_or_ignore(Misc(id=1), session)
         insert_or_ignore(SMTP(id=1), session)
     finally: