+++ mycodo/tests/software_tests/conftest.py
@@ -6,7 +6,7 @@ import tempfile
 import shutil
 import os
 from mycodo.databases.utils import session_scope
-from mycodo.databases.mycodo_db.models_5 import Users
+from mycodo.databases.mycodo_db.models_5 import User
 
 
 def uri_to_path(uri):
@@ -51,9 +51,9 @@ def create_admin_user(mycodo_db_uri):
     """ mycodo_flask exits if there is no user called admin. So we create one """
 
     with session_scope(mycodo_db_uri) as db_session:
-        if not db_session.query(Users).filter_by(user_restriction='admin').count():
+        if not db_session.query(User).filter_by(user_role=1).count():
             logging.info("--> Creating new 'test' user as an admin")
-            db_session.add(Users(user_name='test', user_restriction='admin'))
+            db_session.add(User(user_name='test', user_role=1))
             db_session.commit()
         else:
             logging.warning("--> Dirty User DB: Admin user was already setup in: '{uri}'".format(uri=user_db_uri))