+++ mycodo/tests/software_tests/test_fixtures_and_factories.py
@@ -21,12 +21,12 @@ from mycodo.tests.software_tests.factories_user import (
 # --------------------------
 #   Factory Tests
 # --------------------------
-def create_user_from_factory(factory, user_restriction=None):
+def create_user_from_factory(factory, user_role=None):
     """ Uses a factory to create a user and attempts to save it """
     new_user = factory()
     assert new_user.user_name, "Failed to create a 'user_name' using {factory}".format(factory=factory)
     assert new_user.user_email, "Failed to create a 'user_email' using {factory}".format(factory=factory)
-    assert new_user.user_restriction == user_restriction
+    assert new_user.user_role == user_role
 
 
 def test_user_factories_creates_valid_user():
@@ -36,9 +36,9 @@ def test_user_factories_creates_valid_user():
 
 def test_admin_factories_creates_valid_user():
     """ Use AdminFactory to create new user"""
-    create_user_from_factory(AdminFactory, user_restriction='admin')
+    create_user_from_factory(AdminFactory, user_role=1)
 
 
 def test_guest_factories_creates_valid_user():
     """ Use GuestFactory to create new user"""
-    create_user_from_factory(GuestFactory, user_restriction='guest')
+    create_user_from_factory(GuestFactory, user_role=4)